+++ src/NVP_nanoCAD_COM/OdaX/AcadBlocks.cs
@@ -99,4 +99,26 @@ namespace OdaX.AcadBlocks
 
 		}
 	}
+
+
+    [NVP_Manifest(
+        Text = "Возвращает все Блоки в виде OdaX.AcadBlock",
+        ViewType = "Data")]
+    [NodeInput("AcadBlocks", typeof(object))]
+    public class GetAll_Blocks : INode
+    {
+        public NodeResult Execute(INVPData context, List<NodeResult> inputs)
+        {
+            var _input0 = ((dynamic)inputs[0].Value)._i as OdaX.IAcadBlocks;
+            List<OdaX.AcadBlock.AcadBlock_Constructor> items = new List<OdaX.AcadBlock.AcadBlock_Constructor>();
+            for (int item_counter = 0; item_counter < _input0.Count; item_counter++)
+            {
+				OdaX.AcadBlock.AcadBlock_Constructor item = new AcadBlock.AcadBlock_Constructor();
+                item._i = _input0.Item(item_counter);
+                items.Add(item);
+            }
+            return new NodeResult(items);
+
+        }
+    }
 }