+++ src/NVP_nanoCAD_COM/OdaX/AcadHyperlinks.cs
@@ -119,4 +119,25 @@ namespace OdaX.AcadHyperlinks
 
 		}
 	}
+
+    [NVP_Manifest(
+        Text = "Возвращает все Ссылки в виде OdaX.AcadHyperlink",
+        ViewType = "Data")]
+    [NodeInput("AcadHyperlinks", typeof(object))]
+    public class GetAll_Links : INode
+    {
+        public NodeResult Execute(INVPData context, List<NodeResult> inputs)
+        {
+            var _input0 = ((dynamic)inputs[0].Value)._i as OdaX.IAcadHyperlinks;
+            List<OdaX.AcadHyperlink.AcadHyperlink_Constructor> items = new List<OdaX.AcadHyperlink.AcadHyperlink_Constructor>();
+            for (int item_counter = 0; item_counter < _input0.Count; item_counter++)
+            {
+                OdaX.AcadHyperlink.AcadHyperlink_Constructor item = new OdaX.AcadHyperlink.AcadHyperlink_Constructor();
+                item._i = _input0.Item(item_counter);
+                items.Add(item);
+            }
+            return new NodeResult(items);
+
+        }
+    }
 }