+++ src/NVP_nanoCAD_COM/OdaX/AcadLayouts.cs
@@ -98,4 +98,25 @@ namespace OdaX.AcadLayouts
 
 		}
 	}
+
+    [NVP_Manifest(
+        Text = "Возвращает все Листы в виде OdaX.AcadLayout",
+        ViewType = "Data")]
+    [NodeInput("AcadLayers", typeof(object))]
+    public class GetAll_Layouts : INode
+    {
+        public NodeResult Execute(INVPData context, List<NodeResult> inputs)
+        {
+            var _input0 = ((dynamic)inputs[0].Value)._i as OdaX.IAcadLayouts;
+            List<OdaX.AcadLayout.AcadLayout_Constructor> items = new List<OdaX.AcadLayout.AcadLayout_Constructor>();
+            for (int item_counter = 0; item_counter < _input0.Count; item_counter++)
+            {
+                OdaX.AcadLayout.AcadLayout_Constructor item = new OdaX.AcadLayout.AcadLayout_Constructor();
+                item._i = _input0.Item(item_counter);
+                items.Add(item);
+            }
+            return new NodeResult(items);
+
+        }
+    }
 }