+++ src/NVP_nanoCAD_COM/OdaX/AcadRegisteredApplications.cs
@@ -98,4 +98,26 @@ namespace OdaX.AcadRegisteredApplications
 
 		}
 	}
+
+
+    [NVP_Manifest(
+        Text = "Возвращает все Зарегистрированные приложения в виде OdaX.AcadRegisteredApplication",
+        ViewType = "Data")]
+    [NodeInput("AcadRegisteredApplications", typeof(object))]
+    public class GetAll_RegisteredApplications : INode
+    {
+        public NodeResult Execute(INVPData context, List<NodeResult> inputs)
+        {
+            var _input0 = ((dynamic)inputs[0].Value)._i as OdaX.IAcadRegisteredApplications;
+            List<OdaX.AcadRegisteredApplication.AcadRegisteredApplication_Constructor> items = new List<OdaX.AcadRegisteredApplication.AcadRegisteredApplication_Constructor>();
+            for (int item_counter = 0; item_counter < _input0.Count; item_counter++)
+            {
+                OdaX.AcadRegisteredApplication.AcadRegisteredApplication_Constructor item = new OdaX.AcadRegisteredApplication.AcadRegisteredApplication_Constructor();
+                item._i = _input0.Item(item_counter);
+                items.Add(item);
+            }
+            return new NodeResult(items);
+
+        }
+    }
 }